refactor: test code divide aws and naver
